+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Modal/tests/Modal.test.js
@@ -18,6 +18,29 @@ test('The component should render in body when open', () => {
             <p>My modal content</p>
         </Modal>
     ).render();
+
+    expect(view).toMatchSnapshot();
+    expect(pretty(body.innerHTML)).toMatchSnapshot();
+});
+
+test('The component should render in body with actions when open', () => {
+    const actions = [
+        {title: 'Action 1', onClick: () => {}},
+        {title: 'Action 2', onClick: () => {}},
+    ];
+    const body = document.body;
+    const onRequestClose = () => {};
+    const view = mount(
+        <Modal
+            title="My modal title"
+            onRequestClose={onRequestClose}
+            confirmText="Apply"
+            actions={actions}
+            isOpen={true}>
+            <p>My modal content</p>
+        </Modal>
+    ).render();
+
     expect(view).toMatchSnapshot();
     expect(pretty(body.innerHTML)).toMatchSnapshot();
 });
@@ -55,4 +78,39 @@ test('The component should request to be closed on click on backdrop', () => {
     expect(requestCloseSpy).not.toBeCalled();
     backdrop.props().onClick();
     expect(requestCloseSpy).toBeCalled();
+});
+
+test('The component should request to be closed when the close icon is clicked', () => {
+    const requestCloseSpy = jest.fn();
+    const view = shallow(
+        <Modal
+            title="My modal title"
+            onRequestClose={requestCloseSpy}
+            confirmText="Apply"
+            isOpen={true}>
+            <p>My modal content</p>
+        </Modal>
+    );
+
+    expect(requestCloseSpy).not.toBeCalled();
+    view.find('Icon').simulate('click');
+    expect(requestCloseSpy).toBeCalled();
+});
+
+test('The component should call the callback when the confirm button is clicked', () => {
+    const onRequestClose = () => {};
+    const onConfirm = jest.fn();
+    const view = shallow(
+        <Modal
+            title="My title"
+            onRequestClose={onRequestClose}
+            onConfirm={onConfirm}
+            confirmText="Alright mate!">
+            <p>My modal content</p>
+        </Modal>
+    );
+
+    expect(onConfirm).not.toBeCalled();
+    view.find('.confirmButton').simulate('click');
+    expect(onConfirm).toBeCalled();
 });
