+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Modal/types.js
@@ -1,15 +1,5 @@
 // @flow
-import type {Node} from 'react';
-
 export type Action = {
     title: string,
     onClick: () => void,
 };
-
-export type OverlayProps = {
-    title: string,
-    children: Node,
-    actions: Array<Action>,
-    confirmText: string,
-    onConfirm: () => void,
-}
